rewrite: Refactored enumeration parsing/serialization code to be more reuseable outside of ModelInspector
