+++ src/Hl7.Fhir.Core/Serialization/PrimitiveValueWriter.cs
@@ -41,10 +41,7 @@ namespace Hl7.Fhir.Serialization
 
                 if (nativeType.IsEnum())
                 {
-                    var enumMapping = _inspector.FindEnumMappingByType(nativeType);
-
-                    if (enumMapping != null)
-                        value = enumMapping.GetLiteral((Enum)value);
+                    value = ((Enum)value).GetLiteral();
                 }
             }
